Move CLI source code
