+++ src/modules/cli/midi.test.js
